[rename] MIR -> LIR, HIR -> MIR
Prepare for introducing a new HIR AST set to represent lowered program, but without assumption of optimized enum representation.
