+++ crates/samlang-core/src/compiler/hir_lowering.rs
@@ -1,17 +1,17 @@
 use super::{
-  hir_constant_param_elimination, hir_generics_specialization,
+  hir_generics_specialization,
   hir_string_manager::StringManager,
-  hir_tail_recursion_rewrite,
-  hir_type_conversion::{
+  mir_constant_param_elimination, mir_tail_recursion_rewrite,
+  mir_type_conversion::{
     collect_used_generic_types, type_application, SynthesizedTypes, TypeLoweringManager,
     TypeSynthesizer,
   },
-  hir_type_deduplication,
+  mir_type_deduplication,
 };
 use crate::{
   ast::{
     common_names::{self, encode_samlang_type},
-    hir, source,
+    mir, source,
   },
   checker::type_,
   common::{well_known_pstrs, Heap, LocalStackedContext, ModuleReference, PStr},
@@ -23,25 +23,25 @@ use std::{
 };
 
 struct LoweringResult {
-  statements: Vec<hir::Statement>,
-  expression: hir::Expression,
+  statements: Vec<mir::Statement>,
+  expression: mir::Expression,
 }
 
 struct LoweringResultWithSyntheticFunctions {
-  synthetic_functions: Vec<hir::Function>,
-  statements: Vec<hir::Statement>,
-  expression: hir::Expression,
+  synthetic_functions: Vec<mir::Function>,
+  statements: Vec<mir::Statement>,
+  expression: mir::Expression,
 }
 
-type LoweringContext = LocalStackedContext<PStr, hir::Expression>;
+type LoweringContext = LocalStackedContext<PStr, mir::Expression>;
 
 impl LoweringContext {
-  fn bind(&mut self, name: PStr, value: hir::Expression) {
+  fn bind(&mut self, name: PStr, value: mir::Expression) {
     match &value {
-      hir::Expression::IntLiteral(_) | hir::Expression::Variable(_) => {
+      mir::Expression::IntLiteral(_) | mir::Expression::Variable(_) => {
         self.insert(name, value);
       }
-      hir::Expression::StringName(n) => {
+      mir::Expression::StringName(n) => {
         let value_to_insert = self.get(n).cloned().unwrap_or(value);
         self.insert(name, value_to_insert);
       }
@@ -59,7 +59,7 @@ mod lowering_cx_boilterplate_tests {
 
     LoweringContext::new().bind(
       heap.alloc_str_for_test("a"),
-      hir::Expression::var_name(heap.alloc_str_for_test("a"), hir::INT_TYPE),
+      mir::Expression::var_name(heap.alloc_str_for_test("a"), mir::INT_TYPE),
     );
   }
 }
@@ -69,8 +69,8 @@ struct ExpressionLoweringManager<'a> {
   heap: &'a mut Heap,
   module_reference: &'a ModuleReference,
   encoded_function_name: PStr,
-  defined_variables: Vec<(PStr, hir::Type)>,
-  type_definition_mapping: &'a HashMap<PStr, hir::TypeDefinition>,
+  defined_variables: Vec<(PStr, mir::Type)>,
+  type_definition_mapping: &'a HashMap<PStr, mir::TypeDefinition>,
   type_lowering_manager: &'a mut TypeLoweringManager,
   string_manager: &'a mut StringManager,
   // Mutable states
@@ -79,22 +79,22 @@ struct ExpressionLoweringManager<'a> {
   depth: i32,
   block_id: i32,
   variable_cx: LoweringContext,
-  synthetic_functions: Vec<hir::Function>,
+  synthetic_functions: Vec<mir::Function>,
 }
 
 impl<'a> ExpressionLoweringManager<'a> {
   fn new(
     module_reference: &'a ModuleReference,
     encoded_function_name: PStr,
-    defined_variables: Vec<(PStr, hir::Type)>,
-    type_definition_mapping: &'a HashMap<PStr, hir::TypeDefinition>,
+    defined_variables: Vec<(PStr, mir::Type)>,
+    type_definition_mapping: &'a HashMap<PStr, mir::TypeDefinition>,
     heap: &'a mut Heap,
     type_lowering_manager: &'a mut TypeLoweringManager,
     string_manager: &'a mut StringManager,
   ) -> ExpressionLoweringManager<'a> {
     let mut variable_cx = LoweringContext::new();
     for (n, t) in &defined_variables {
-      variable_cx.bind(*n, hir::Expression::var_name(*n, t.clone()));
+      variable_cx.bind(*n, mir::Expression::var_name(*n, t.clone()));
     }
     ExpressionLoweringManager {
       heap,
@@ -136,52 +136,52 @@ impl<'a> ExpressionLoweringManager<'a> {
     &mut self,
     expression: &source::expr::E<Rc<type_::Type>>,
     favored_temp_variable: Option<PStr>,
-    statements: &mut Vec<hir::Statement>,
-  ) -> hir::Expression {
+    statements: &mut Vec<mir::Statement>,
+  ) -> mir::Expression {
     let LoweringResult { statements: mut lowered_statements, expression: e } =
       self.lower(expression, favored_temp_variable);
     statements.append(&mut lowered_statements);
     e
   }
 
-  fn get_synthetic_identifier_type_from_tuple(&mut self, mappings: Vec<hir::Type>) -> hir::IdType {
+  fn get_synthetic_identifier_type_from_tuple(&mut self, mappings: Vec<mir::Type>) -> mir::IdType {
     let type_parameters = collect_used_generic_types(
-      &hir::Type::new_fn_unwrapped(mappings.clone(), hir::INT_TYPE),
+      &mir::Type::new_fn_unwrapped(mappings.clone(), mir::INT_TYPE),
       &self.type_lowering_manager.generic_types,
     )
     .into_iter()
     .sorted()
     .collect_vec();
     let type_arguments =
-      type_parameters.iter().cloned().map(hir::Type::new_id_no_targs).collect_vec();
+      type_parameters.iter().cloned().map(mir::Type::new_id_no_targs).collect_vec();
     let name = self
       .type_lowering_manager
       .type_synthesizer
       .synthesize_tuple_type(self.heap, mappings, type_parameters)
       .identifier;
-    hir::IdType { name, type_arguments }
+    mir::IdType { name, type_arguments }
   }
 
   fn get_synthetic_identifier_type_from_closure(
     &mut self,
-    fn_type: hir::FunctionType,
-  ) -> hir::IdType {
+    fn_type: mir::FunctionType,
+  ) -> mir::IdType {
     let type_parameters =
       collect_used_generic_types(&fn_type, &self.type_lowering_manager.generic_types)
         .into_iter()
         .sorted()
         .collect_vec();
     let type_arguments =
-      type_parameters.iter().cloned().map(hir::Type::new_id_no_targs).collect_vec();
+      type_parameters.iter().cloned().map(mir::Type::new_id_no_targs).collect_vec();
     let name = self
       .type_lowering_manager
       .type_synthesizer
       .synthesize_closure_type(self.heap, fn_type, type_parameters)
       .identifier;
-    hir::IdType { name, type_arguments }
+    mir::IdType { name, type_arguments }
   }
 
-  fn resolve_variable(&mut self, variable_name: &PStr) -> hir::Expression {
+  fn resolve_variable(&mut self, variable_name: &PStr) -> mir::Expression {
     self
       .variable_cx
       .get(variable_name)
@@ -189,7 +189,7 @@ impl<'a> ExpressionLoweringManager<'a> {
       .clone()
   }
 
-  fn resolve_struct_mapping_of_id_type(&mut self, hir_id_type: &hir::IdType) -> Vec<hir::Type> {
+  fn resolve_struct_mapping_of_id_type(&mut self, hir_id_type: &mir::IdType) -> Vec<mir::Type> {
     let type_def = self.type_definition_mapping.get(&hir_id_type.name).unwrap().clone();
     let replacement_map: HashMap<_, _> =
       type_def.type_parameters.iter().cloned().zip(hir_id_type.type_arguments.clone()).collect();
@@ -202,7 +202,7 @@ impl<'a> ExpressionLoweringManager<'a> {
       .collect()
   }
 
-  fn get_function_type_without_context(&mut self, t: &type_::Type) -> hir::FunctionType {
+  fn get_function_type_without_context(&mut self, t: &type_::Type) -> mir::FunctionType {
     let type_::FunctionType { argument_types, return_type, .. } =
       t.as_fn().expect("Expecting function type");
     let (_, t) = self.type_lowering_manager.lower_source_function_type_for_toplevel(
@@ -220,14 +220,14 @@ impl<'a> ExpressionLoweringManager<'a> {
   ) -> LoweringResult {
     match expression {
       source::expr::E::Literal(_, source::Literal::Bool(b)) => {
-        LoweringResult { statements: vec![], expression: if *b { hir::ONE } else { hir::ZERO } }
+        LoweringResult { statements: vec![], expression: if *b { mir::ONE } else { mir::ZERO } }
       }
       source::expr::E::Literal(_, source::Literal::Int(i)) => {
-        LoweringResult { statements: vec![], expression: hir::Expression::int(*i) }
+        LoweringResult { statements: vec![], expression: mir::Expression::int(*i) }
       }
       source::expr::E::Literal(_, source::Literal::String(s)) => LoweringResult {
         statements: vec![],
-        expression: hir::Expression::StringName(self.string_manager.allocate(self.heap, *s).name),
+        expression: mir::Expression::StringName(self.string_manager.allocate(self.heap, *s).name),
       },
       source::expr::E::LocalId(_, id) => LoweringResult {
         statements: vec![],
@@ -238,7 +238,7 @@ impl<'a> ExpressionLoweringManager<'a> {
         },
       },
       source::expr::E::ClassId(_, _, _) => {
-        LoweringResult { statements: vec![], expression: hir::ZERO }
+        LoweringResult { statements: vec![], expression: mir::ZERO }
       }
       source::expr::E::FieldAccess(e) => self.lower_field_access(e, favored_temp_variable),
       source::expr::E::MethodAccess(e) => self.lower_method_access(e, favored_temp_variable),
@@ -284,7 +284,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     let index = usize::try_from(expression.field_order).unwrap();
     let extracted_field_type = &mappings_for_id_type[index];
     let value_name = self.allocate_temp_variable(favored_temp_variable);
-    statements.push(hir::Statement::IndexedAccess {
+    statements.push(mir::Statement::IndexedAccess {
       name: value_name,
       type_: extracted_field_type.clone(),
       pointer_expression: result_expr,
@@ -292,10 +292,10 @@ impl<'a> ExpressionLoweringManager<'a> {
     });
     self
       .variable_cx
-      .bind(value_name, hir::Expression::var_name(value_name, extracted_field_type.clone()));
+      .bind(value_name, mir::Expression::var_name(value_name, extracted_field_type.clone()));
     LoweringResult {
       statements,
-      expression: hir::Expression::var_name(value_name, extracted_field_type.clone()),
+      expression: mir::Expression::var_name(value_name, extracted_field_type.clone()),
     }
   }
 
@@ -312,7 +312,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     let LoweringResult { mut statements, expression: result_expr } =
       self.lower(&expression.object, None);
     let original_function_type = self.get_function_type_without_context(&expression.common.type_);
-    let method_type = hir::FunctionType {
+    let method_type = mir::FunctionType {
       argument_types: vec![result_expr.type_().clone()]
         .into_iter()
         .chain(original_function_type.argument_types.iter().cloned())
@@ -323,12 +323,12 @@ impl<'a> ExpressionLoweringManager<'a> {
     let closure_variable_name = self.allocate_temp_variable(favored_temp_variable);
     self.variable_cx.bind(
       closure_variable_name,
-      hir::Expression::var_name(closure_variable_name, hir::Type::Id(closure_type.clone())),
+      mir::Expression::var_name(closure_variable_name, mir::Type::Id(closure_type.clone())),
     );
-    statements.push(hir::Statement::ClosureInit {
+    statements.push(mir::Statement::ClosureInit {
       closure_variable_name,
       closure_type: closure_type.clone(),
-      function_name: hir::FunctionName {
+      function_name: mir::FunctionName {
         name: self.heap.alloc_string(function_name),
         type_: method_type,
         type_arguments: self
@@ -339,7 +339,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     });
     LoweringResult {
       statements,
-      expression: hir::Expression::var_name(closure_variable_name, hir::Type::Id(closure_type)),
+      expression: mir::Expression::var_name(closure_variable_name, mir::Type::Id(closure_type)),
     }
   }
 
@@ -352,21 +352,21 @@ impl<'a> ExpressionLoweringManager<'a> {
       self.lower(&expression.argument, None);
     let value_name = self.allocate_temp_variable(favored_temp_variable);
     let new_binary = match expression.operator {
-      source::expr::UnaryOperator::NOT => hir::Binary {
+      source::expr::UnaryOperator::NOT => mir::Binary {
         name: value_name,
-        operator: hir::Operator::XOR,
+        operator: mir::Operator::XOR,
         e1: result_expr,
-        e2: hir::ONE,
+        e2: mir::ONE,
       },
-      source::expr::UnaryOperator::NEG => hir::Binary {
+      source::expr::UnaryOperator::NEG => mir::Binary {
         name: value_name,
-        operator: hir::Operator::MINUS,
-        e1: hir::ZERO,
+        operator: mir::Operator::MINUS,
+        e1: mir::ZERO,
         e2: result_expr,
       },
     };
-    let final_expr = hir::Expression::var_name(new_binary.name, hir::INT_TYPE);
-    statements.push(hir::Statement::Binary(new_binary));
+    let final_expr = mir::Expression::var_name(new_binary.name, mir::INT_TYPE);
+    statements.push(mir::Statement::Binary(new_binary));
     LoweringResult { statements, expression: final_expr }
   }
 
@@ -404,10 +404,10 @@ impl<'a> ExpressionLoweringManager<'a> {
         };
         (
           fn_type_without_cx.return_type.as_ref().clone(),
-          hir::Statement::Call {
-            callee: hir::Callee::FunctionName(hir::FunctionName {
+          mir::Statement::Call {
+            callee: mir::Callee::FunctionName(mir::FunctionName {
               name: self.heap.alloc_string(fn_name),
-              type_: hir::FunctionType {
+              type_: mir::FunctionType {
                 argument_types: vec![hir_target.type_().clone()]
                   .into_iter()
                   .chain(fn_type_without_cx.argument_types.iter().cloned())
@@ -448,8 +448,8 @@ impl<'a> ExpressionLoweringManager<'a> {
           .collect_vec();
         (
           return_type.clone(),
-          hir::Statement::Call {
-            callee: hir::Callee::Variable(lowered_fn_expr),
+          mir::Statement::Call {
+            callee: mir::Callee::Variable(lowered_fn_expr),
             arguments: lowered_args,
             return_type,
             return_collector: if is_void_return { None } else { Some(return_collector_name) },
@@ -462,9 +462,9 @@ impl<'a> ExpressionLoweringManager<'a> {
     LoweringResult {
       statements: lowered_stmts,
       expression: if is_void_return {
-        hir::ZERO
+        mir::ZERO
       } else {
-        hir::Expression::var_name(return_collector_name, function_return_collector_type)
+        mir::Expression::var_name(return_collector_name, function_return_collector_type)
       },
     }
   }
@@ -485,26 +485,26 @@ impl<'a> ExpressionLoweringManager<'a> {
           self.lower_binary(&expression.e1, None);
         let LoweringResult { statements: s2, expression: e2 } =
           self.lower_binary(&expression.e2, None);
-        if let hir::Expression::IntLiteral(v) = &e1 {
+        if let mir::Expression::IntLiteral(v) = &e1 {
           return if *v != 0 {
             LoweringResult {
               statements: s1.into_iter().chain(s2.into_iter()).collect_vec(),
               expression: e2,
             }
           } else {
-            LoweringResult { statements: s1, expression: hir::ZERO }
+            LoweringResult { statements: s1, expression: mir::ZERO }
           };
         }
         let mut statements = s1;
-        statements.push(hir::Statement::IfElse {
+        statements.push(mir::Statement::IfElse {
           condition: e1,
           s1: s2,
           s2: vec![],
-          final_assignments: vec![(temp, hir::INT_TYPE, e2, hir::ZERO)],
+          final_assignments: vec![(temp, mir::INT_TYPE, e2, mir::ZERO)],
         });
         return LoweringResult {
           statements,
-          expression: hir::Expression::var_name(temp, hir::INT_TYPE),
+          expression: mir::Expression::var_name(temp, mir::INT_TYPE),
         };
       }
       source::expr::BinaryOperator::OR => {
@@ -513,9 +513,9 @@ impl<'a> ExpressionLoweringManager<'a> {
           self.lower_binary(&expression.e1, None);
         let LoweringResult { statements: s2, expression: e2 } =
           self.lower_binary(&expression.e2, None);
-        if let hir::Expression::IntLiteral(v) = &e1 {
+        if let mir::Expression::IntLiteral(v) = &e1 {
           return if *v != 0 {
-            LoweringResult { statements: s1, expression: hir::ONE }
+            LoweringResult { statements: s1, expression: mir::ONE }
           } else {
             LoweringResult {
               statements: s1.into_iter().chain(s2.into_iter()).collect_vec(),
@@ -524,15 +524,15 @@ impl<'a> ExpressionLoweringManager<'a> {
           };
         }
         let mut statements = s1;
-        statements.push(hir::Statement::IfElse {
+        statements.push(mir::Statement::IfElse {
           condition: e1,
           s1: vec![],
           s2,
-          final_assignments: vec![(temp, hir::INT_TYPE, hir::ONE, e2)],
+          final_assignments: vec![(temp, mir::INT_TYPE, mir::ONE, e2)],
         });
         return LoweringResult {
           statements,
-          expression: hir::Expression::var_name(temp, hir::INT_TYPE),
+          expression: mir::Expression::var_name(temp, mir::INT_TYPE),
         };
       }
       source::expr::BinaryOperator::CONCAT => {
@@ -545,7 +545,7 @@ impl<'a> ExpressionLoweringManager<'a> {
           let concat_pstr = self.heap.alloc_string(concat_string);
           return LoweringResult {
             statements: vec![],
-            expression: hir::Expression::StringName(
+            expression: mir::Expression::StringName(
               self.string_manager.allocate(self.heap, concat_pstr).name,
             ),
           };
@@ -554,44 +554,44 @@ impl<'a> ExpressionLoweringManager<'a> {
         let e1 = self.lowered_and_add_statements(&expression.e1, None, &mut lowered_stmts);
         let e2 = self.lowered_and_add_statements(&expression.e2, None, &mut lowered_stmts);
         let return_collector_name = self.allocate_temp_variable(favored_temp_variable);
-        lowered_stmts.push(hir::Statement::Call {
-          callee: hir::Callee::FunctionName(hir::FunctionName {
+        lowered_stmts.push(mir::Statement::Call {
+          callee: mir::Callee::FunctionName(mir::FunctionName {
             name: self.heap.alloc_string(common_names::encoded_fn_name_string_concat()),
-            type_: hir::Type::new_fn_unwrapped(
-              vec![hir::STRING_TYPE, hir::STRING_TYPE],
-              hir::STRING_TYPE,
+            type_: mir::Type::new_fn_unwrapped(
+              vec![mir::STRING_TYPE, mir::STRING_TYPE],
+              mir::STRING_TYPE,
             ),
             type_arguments: vec![],
           }),
           arguments: vec![e1, e2],
-          return_type: hir::STRING_TYPE,
+          return_type: mir::STRING_TYPE,
           return_collector: Some(return_collector_name),
         });
         return LoweringResult {
           statements: lowered_stmts,
-          expression: hir::Expression::var_name(return_collector_name, hir::STRING_TYPE),
+          expression: mir::Expression::var_name(return_collector_name, mir::STRING_TYPE),
         };
       }
-      source::expr::BinaryOperator::MUL => hir::Operator::MUL,
-      source::expr::BinaryOperator::DIV => hir::Operator::DIV,
-      source::expr::BinaryOperator::MOD => hir::Operator::MOD,
-      source::expr::BinaryOperator::PLUS => hir::Operator::PLUS,
-      source::expr::BinaryOperator::MINUS => hir::Operator::MINUS,
-      source::expr::BinaryOperator::LT => hir::Operator::LT,
-      source::expr::BinaryOperator::LE => hir::Operator::LE,
-      source::expr::BinaryOperator::GT => hir::Operator::GT,
-      source::expr::BinaryOperator::GE => hir::Operator::GE,
-      source::expr::BinaryOperator::EQ => hir::Operator::EQ,
-      source::expr::BinaryOperator::NE => hir::Operator::NE,
+      source::expr::BinaryOperator::MUL => mir::Operator::MUL,
+      source::expr::BinaryOperator::DIV => mir::Operator::DIV,
+      source::expr::BinaryOperator::MOD => mir::Operator::MOD,
+      source::expr::BinaryOperator::PLUS => mir::Operator::PLUS,
+      source::expr::BinaryOperator::MINUS => mir::Operator::MINUS,
+      source::expr::BinaryOperator::LT => mir::Operator::LT,
+      source::expr::BinaryOperator::LE => mir::Operator::LE,
+      source::expr::BinaryOperator::GT => mir::Operator::GT,
+      source::expr::BinaryOperator::GE => mir::Operator::GE,
+      source::expr::BinaryOperator::EQ => mir::Operator::EQ,
+      source::expr::BinaryOperator::NE => mir::Operator::NE,
     };
     let mut lowered_stmts = vec![];
     let e1 = self.lowered_and_add_statements(&expression.e1, None, &mut lowered_stmts);
     let e2 = self.lowered_and_add_statements(&expression.e2, None, &mut lowered_stmts);
     let value_temp = self.allocate_temp_variable(favored_temp_variable);
-    lowered_stmts.push(hir::Statement::binary(value_temp, operator, e1, e2));
+    lowered_stmts.push(mir::Statement::binary(value_temp, operator, e1, e2));
     LoweringResult {
       statements: lowered_stmts,
-      expression: hir::Expression::var_name(value_temp, hir::INT_TYPE),
+      expression: mir::Expression::var_name(value_temp, mir::INT_TYPE),
     }
   }
 
@@ -607,7 +607,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     let LoweringResult { statements: s1, expression: e1 } = self.lower(&expression.e1, None);
     let LoweringResult { statements: s2, expression: e2 } = self.lower(&expression.e2, None);
     let lowered_return_type = e1.type_().clone();
-    lowered_stmts.push(hir::Statement::IfElse {
+    lowered_stmts.push(mir::Statement::IfElse {
       condition,
       s1,
       s2,
@@ -615,10 +615,10 @@ impl<'a> ExpressionLoweringManager<'a> {
     });
     self
       .variable_cx
-      .bind(final_var_name, hir::Expression::var_name(final_var_name, lowered_return_type.clone()));
+      .bind(final_var_name, mir::Expression::var_name(final_var_name, lowered_return_type.clone()));
     LoweringResult {
       statements: lowered_stmts,
-      expression: hir::Expression::var_name(final_var_name, lowered_return_type),
+      expression: mir::Expression::var_name(final_var_name, lowered_return_type),
     }
   }
 
@@ -628,15 +628,15 @@ impl<'a> ExpressionLoweringManager<'a> {
       self.lowered_and_add_statements(&expression.matched, None, &mut lowered_stmts);
     let general_id_type = matched_expr.type_().as_id().unwrap();
     let variable_for_tag = self.allocate_temp_variable(None);
-    lowered_stmts.push(hir::Statement::IndexedAccess {
+    lowered_stmts.push(mir::Statement::IndexedAccess {
       name: variable_for_tag,
-      type_: hir::INT_TYPE,
+      type_: mir::INT_TYPE,
       pointer_expression: matched_expr.clone(),
       index: 0,
     });
     self
       .variable_cx
-      .bind(variable_for_tag, hir::Expression::var_name(variable_for_tag, hir::INT_TYPE));
+      .bind(variable_for_tag, mir::Expression::var_name(variable_for_tag, mir::INT_TYPE));
 
     let mut lowered_matching_list = vec![];
     for source::expr::VariantPatternToExpression {
@@ -649,7 +649,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     {
       let mut local_stmts = vec![];
       let casted_collector = self.allocate_temp_variable(None);
-      let subtype = hir::Type::Id(hir::IdType {
+      let subtype = mir::Type::Id(mir::IdType {
         name: self.heap.alloc_string(format!(
           "{}_{}",
           general_id_type.name.as_str(self.heap),
@@ -657,7 +657,7 @@ impl<'a> ExpressionLoweringManager<'a> {
         )),
         type_arguments: general_id_type.type_arguments.clone(),
       });
-      local_stmts.push(hir::Statement::Cast {
+      local_stmts.push(mir::Statement::Cast {
         name: casted_collector,
         type_: subtype.clone(),
         assigned_expression: matched_expr.clone(),
@@ -668,13 +668,13 @@ impl<'a> ExpressionLoweringManager<'a> {
       {
         let name = data_var_name.name;
         let data_var_type = self.type_lowering_manager.lower_source_type(self.heap, data_var_type);
-        local_stmts.push(hir::Statement::IndexedAccess {
+        local_stmts.push(mir::Statement::IndexedAccess {
           name,
           type_: data_var_type.clone(),
-          pointer_expression: hir::Expression::var_name(casted_collector, subtype.clone()),
+          pointer_expression: mir::Expression::var_name(casted_collector, subtype.clone()),
           index: i + 1,
         });
-        self.variable_cx.bind(name, hir::Expression::var_name(name, data_var_type.clone()));
+        self.variable_cx.bind(name, mir::Expression::var_name(name, data_var_type.clone()));
       }
       let final_expr = self.lowered_and_add_statements(body, None, &mut local_stmts);
       self.variable_cx.pop_scope();
@@ -690,14 +690,14 @@ impl<'a> ExpressionLoweringManager<'a> {
       let lowered_return_type = acc.1.type_().clone();
       let (acc_stmts, acc_e) = acc;
       let new_stmts = vec![
-        hir::Statement::Binary(hir::Statement::binary_unwrapped(
+        mir::Statement::Binary(mir::Statement::binary_unwrapped(
           comparison_temp,
-          hir::Operator::EQ,
-          hir::Expression::var_name(variable_for_tag, hir::INT_TYPE),
-          hir::Expression::int(i32::try_from(tag_order).unwrap()),
+          mir::Operator::EQ,
+          mir::Expression::var_name(variable_for_tag, mir::INT_TYPE),
+          mir::Expression::int(i32::try_from(tag_order).unwrap()),
         )),
-        hir::Statement::IfElse {
-          condition: hir::Expression::var_name(comparison_temp, hir::INT_TYPE),
+        mir::Statement::IfElse {
+          condition: mir::Expression::var_name(comparison_temp, mir::INT_TYPE),
           s1: case_stmts,
           s2: acc_stmts,
           final_assignments: vec![(
@@ -708,7 +708,7 @@ impl<'a> ExpressionLoweringManager<'a> {
           )],
         },
       ];
-      acc = (new_stmts, hir::Expression::var_name(final_assignment_temp, lowered_return_type))
+      acc = (new_stmts, mir::Expression::var_name(final_assignment_temp, lowered_return_type))
     }
 
     lowered_stmts.append(&mut acc.0);
@@ -718,15 +718,15 @@ impl<'a> ExpressionLoweringManager<'a> {
   fn create_synthetic_lambda_function(
     &mut self,
     expression: &source::expr::Lambda<Rc<type_::Type>>,
-    captured: &[(PStr, hir::Expression)],
-    context_type: &hir::Type,
-  ) -> hir::Function {
+    captured: &[(PStr, mir::Expression)],
+    context_type: &mir::Type,
+  ) -> mir::Function {
     let mut lambda_stmts = vec![];
     for (index, (name, e)) in captured.iter().enumerate() {
-      lambda_stmts.push(hir::Statement::IndexedAccess {
+      lambda_stmts.push(mir::Statement::IndexedAccess {
         name: *name,
         type_: e.type_().clone(),
-        pointer_expression: hir::Expression::var_name(
+        pointer_expression: mir::Expression::var_name(
           well_known_pstrs::UNDERSCORE_THIS,
           context_type.clone(),
         ),
@@ -738,7 +738,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     let source_fn_type = expression.common.type_.as_fn().unwrap();
     let (
       type_parameters,
-      hir::FunctionType {
+      mir::FunctionType {
         argument_types: fun_type_without_cx_argument_types,
         return_type: fun_type_without_cx_return_type,
       },
@@ -766,14 +766,14 @@ impl<'a> ExpressionLoweringManager<'a> {
       .lower(&expression.body, None);
     lambda_stmts.append(&mut lowered_s);
 
-    hir::Function {
+    mir::Function {
       name: fn_name,
       parameters: vec![well_known_pstrs::UNDERSCORE_THIS]
         .into_iter()
         .chain(expression.parameters.iter().map(|it| it.name.name))
         .collect_vec(),
       type_parameters,
-      type_: hir::FunctionType {
+      type_: mir::FunctionType {
         argument_types: vec![context_type.clone()]
           .into_iter()
           .chain(fun_type_without_cx_argument_types.into_iter())
@@ -795,40 +795,40 @@ impl<'a> ExpressionLoweringManager<'a> {
     let mut lowered_stmts = vec![];
     let closure_variable_name = self.allocate_temp_variable(favored_temp_variable);
     let context = if captured.is_empty() {
-      hir::ZERO
+      mir::ZERO
     } else {
       let context_name = self.allocate_temp_variable(None);
       let context_type = self.get_synthetic_identifier_type_from_tuple(
         captured.iter().map(|(_, v)| v.type_().clone()).collect_vec(),
       );
-      lowered_stmts.push(hir::Statement::StructInit {
+      lowered_stmts.push(mir::Statement::StructInit {
         struct_variable_name: context_name,
         type_: context_type.clone(),
         expression_list: captured.iter().map(|(_, v)| v.clone()).collect_vec(),
       });
       self.variable_cx.bind(
         context_name,
-        hir::Expression::var_name(context_name, hir::Type::Id(context_type.clone())),
+        mir::Expression::var_name(context_name, mir::Type::Id(context_type.clone())),
       );
-      hir::Expression::var_name(context_name, hir::Type::Id(context_type))
+      mir::Expression::var_name(context_name, mir::Type::Id(context_type))
     };
     let synthetic_lambda =
       self.create_synthetic_lambda_function(expression, &captured, context.type_());
-    let closure_type = self.get_synthetic_identifier_type_from_closure(hir::FunctionType {
+    let closure_type = self.get_synthetic_identifier_type_from_closure(mir::FunctionType {
       argument_types: synthetic_lambda.type_.argument_types.iter().skip(1).cloned().collect_vec(),
       return_type: synthetic_lambda.type_.return_type.clone(),
     });
-    lowered_stmts.push(hir::Statement::ClosureInit {
+    lowered_stmts.push(mir::Statement::ClosureInit {
       closure_variable_name,
       closure_type: closure_type.clone(),
-      function_name: hir::FunctionName {
+      function_name: mir::FunctionName {
         name: synthetic_lambda.name,
         type_: synthetic_lambda.type_.clone(),
         type_arguments: synthetic_lambda
           .type_parameters
           .iter()
           .cloned()
-          .map(hir::Type::new_id_no_targs)
+          .map(mir::Type::new_id_no_targs)
           .collect_vec(),
       },
       context,
@@ -836,15 +836,15 @@ impl<'a> ExpressionLoweringManager<'a> {
     self.synthetic_functions.push(synthetic_lambda);
     self.variable_cx.bind(
       closure_variable_name,
-      hir::Expression::var_name(closure_variable_name, hir::Type::Id(closure_type.clone())),
+      mir::Expression::var_name(closure_variable_name, mir::Type::Id(closure_type.clone())),
     );
     LoweringResult {
       statements: lowered_stmts,
-      expression: hir::Expression::var_name(closure_variable_name, hir::Type::Id(closure_type)),
+      expression: mir::Expression::var_name(closure_variable_name, mir::Type::Id(closure_type)),
     }
   }
 
-  fn get_renamed_variable_for_nesting(&mut self, name: PStr, type_: &hir::Type) -> PStr {
+  fn get_renamed_variable_for_nesting(&mut self, name: PStr, type_: &mir::Type) -> PStr {
     if self.depth == 0 {
       return name;
     }
@@ -854,7 +854,7 @@ impl<'a> ExpressionLoweringManager<'a> {
       self.depth,
       self.block_id
     ));
-    self.variable_cx.bind(name, hir::Expression::var_name(renamed, type_.clone()));
+    self.variable_cx.bind(name, mir::Expression::var_name(renamed, type_.clone()));
     renamed
   }
 
@@ -885,8 +885,8 @@ impl<'a> ExpressionLoweringManager<'a> {
             );
             self
               .variable_cx
-              .bind(mangled_name, hir::Expression::var_name(mangled_name, field_type.clone()));
-            lowered_stmts.push(hir::Statement::IndexedAccess {
+              .bind(mangled_name, mir::Expression::var_name(mangled_name, field_type.clone()));
+            lowered_stmts.push(mir::Statement::IndexedAccess {
               name: mangled_name,
               type_: field_type.clone(),
               pointer_expression: assigned_expr.clone(),
@@ -907,7 +907,7 @@ impl<'a> ExpressionLoweringManager<'a> {
     let final_expr = if let Some(e) = &expression.expression {
       self.lowered_and_add_statements(e, favored_temp_variable, &mut lowered_stmts)
     } else {
-      hir::ZERO
+      mir::ZERO
     };
     self.variable_cx.pop_scope();
     self.block_id += 1;
@@ -935,25 +935,25 @@ fn lower_constructors(
   heap: &mut Heap,
   module_reference: &ModuleReference,
   class_name: PStr,
-  type_definition_mapping: &HashMap<PStr, hir::TypeDefinition>,
-) -> Vec<hir::Function> {
+  type_definition_mapping: &HashMap<PStr, mir::TypeDefinition>,
+) -> Vec<mir::Function> {
   let type_name =
     heap.alloc_string(common_names::encode_samlang_type(heap, module_reference, class_name));
   let type_def = type_definition_mapping.get(&type_name).unwrap();
   let struct_var_name = well_known_pstrs::LOWER_O;
   let struct_var_name_casted = well_known_pstrs::UNDERSCORE_TMP;
-  let struct_type = hir::IdType {
+  let struct_type = mir::IdType {
     name: type_name,
     type_arguments: type_def
       .type_parameters
       .iter()
-      .map(|n| hir::Type::new_id_no_targs(*n))
+      .map(|n| mir::Type::new_id_no_targs(*n))
       .collect_vec(),
   };
   let mut functions = vec![];
   match &type_def.mappings {
-    hir::TypeDefinitionMappings::Struct(types) => {
-      let f = hir::Function {
+    mir::TypeDefinitionMappings::Struct(types) => {
+      let f = mir::Function {
         name: heap.alloc_string(common_names::encode_function_name_globally(
           heap,
           module_reference,
@@ -965,26 +965,26 @@ fn lower_constructors(
           .chain(types.iter().enumerate().map(|(i, _)| heap.alloc_string(format!("_f{i}"))))
           .collect_vec(),
         type_parameters: type_def.type_parameters.clone(),
-        type_: hir::Type::new_fn_unwrapped(
-          vec![hir::INT_TYPE].into_iter().chain(types.iter().cloned()).collect_vec(),
-          hir::Type::Id(struct_type.clone()),
+        type_: mir::Type::new_fn_unwrapped(
+          vec![mir::INT_TYPE].into_iter().chain(types.iter().cloned()).collect_vec(),
+          mir::Type::Id(struct_type.clone()),
         ),
-        body: vec![hir::Statement::StructInit {
+        body: vec![mir::Statement::StructInit {
           struct_variable_name: struct_var_name,
           type_: struct_type.clone(),
           expression_list: types
             .iter()
             .enumerate()
             .map(|(order, t)| {
-              hir::Expression::var_name(heap.alloc_string(format!("_f{order}")), t.clone())
+              mir::Expression::var_name(heap.alloc_string(format!("_f{order}")), t.clone())
             })
             .collect_vec(),
         }],
-        return_value: hir::Expression::var_name(struct_var_name, hir::Type::Id(struct_type)),
+        return_value: mir::Expression::var_name(struct_var_name, mir::Type::Id(struct_type)),
       };
       functions.push(f);
     }
-    hir::TypeDefinitionMappings::Enum => {
+    mir::TypeDefinitionMappings::Enum => {
       for (tag_order, tag_name) in type_def.names.iter().enumerate() {
         let enum_subtype_name = heap.alloc_string(common_names::encode_samlang_variant_subtype(
           heap,
@@ -995,11 +995,11 @@ fn lower_constructors(
         let data_types =
           &type_definition_mapping.get(&enum_subtype_name).unwrap().mappings.as_struct().unwrap()
             [1..];
-        let enum_subtype = hir::IdType {
+        let enum_subtype = mir::IdType {
           name: enum_subtype_name,
           type_arguments: struct_type.type_arguments.clone(),
         };
-        let f = hir::Function {
+        let f = mir::Function {
           name: heap.alloc_string(common_names::encode_function_name_globally(
             heap,
             module_reference,
@@ -1011,36 +1011,36 @@ fn lower_constructors(
             .chain((0..(data_types.len())).map(|i| heap.alloc_string(format!("_data{i}"))))
             .collect(),
           type_parameters: type_def.type_parameters.clone(),
-          type_: hir::Type::new_fn_unwrapped(
-            vec![hir::INT_TYPE].into_iter().chain(data_types.iter().cloned()).collect(),
-            hir::Type::Id(struct_type.clone()),
+          type_: mir::Type::new_fn_unwrapped(
+            vec![mir::INT_TYPE].into_iter().chain(data_types.iter().cloned()).collect(),
+            mir::Type::Id(struct_type.clone()),
           ),
           body: vec![
-            hir::Statement::StructInit {
+            mir::Statement::StructInit {
               struct_variable_name: struct_var_name,
               type_: enum_subtype.clone(),
-              expression_list: vec![hir::Expression::int(i32::try_from(tag_order).unwrap())]
+              expression_list: vec![mir::Expression::int(i32::try_from(tag_order).unwrap())]
                 .into_iter()
                 .chain(data_types.iter().enumerate().map(|(i, data_type)| {
-                  hir::Expression::var_name(
+                  mir::Expression::var_name(
                     heap.alloc_string(format!("_data{i}")),
                     data_type.clone(),
                   )
                 }))
                 .collect(),
             },
-            hir::Statement::Cast {
+            mir::Statement::Cast {
               name: struct_var_name_casted,
-              type_: hir::Type::Id(struct_type.clone()),
-              assigned_expression: hir::Expression::var_name(
+              type_: mir::Type::Id(struct_type.clone()),
+              assigned_expression: mir::Expression::var_name(
                 struct_var_name,
-                hir::Type::Id(enum_subtype),
+                mir::Type::Id(enum_subtype),
               ),
             },
           ],
-          return_value: hir::Expression::var_name(
+          return_value: mir::Expression::var_name(
             struct_var_name_casted,
-            hir::Type::Id(struct_type.clone()),
+            mir::Type::Id(struct_type.clone()),
           ),
         };
         functions.push(f);
@@ -1057,7 +1057,7 @@ fn lower_tparams(type_parameters: &[source::TypeParameter]) -> Vec<PStr> {
 fn compile_sources_with_generics_preserved(
   heap: &mut Heap,
   sources: &HashMap<ModuleReference, source::Module<Rc<type_::Type>>>,
-) -> hir::Sources {
+) -> mir::Sources {
   let mut type_lowering_manager =
     TypeLoweringManager { generic_types: HashSet::new(), type_synthesizer: TypeSynthesizer::new() };
   let mut compiled_type_defs = vec![];
@@ -1119,9 +1119,9 @@ fn compile_sources_with_generics_preserved(
             type_lowering_manager.generic_types = tparams_set;
             let main_function_parameter_with_types = vec![(
               well_known_pstrs::UNDERSCORE_THIS,
-              hir::Type::new_id(
+              mir::Type::new_id(
                 heap.alloc_string(encode_samlang_type(heap, module_reference, c.name.name)),
-                class_tparams.into_iter().map(hir::Type::new_id_no_targs).collect_vec(),
+                class_tparams.into_iter().map(mir::Type::new_id_no_targs).collect_vec(),
               ),
             )]
             .into_iter()
@@ -1147,14 +1147,14 @@ fn compile_sources_with_generics_preserved(
               expression,
               synthetic_functions: mut compiled_functions_to_add,
             } = lower_source_expression(manager, &member.body);
-            let main_fn_type = hir::Type::new_fn_unwrapped(
+            let main_fn_type = mir::Type::new_fn_unwrapped(
               main_function_parameter_with_types.iter().map(|(_, t)| t.clone()).collect_vec(),
               type_lowering_manager.lower_source_type(
                 heap,
                 &type_::Type::from_annotation(&member.decl.type_.return_type),
               ),
             );
-            compiled_functions_to_add.push(hir::Function {
+            compiled_functions_to_add.push(mir::Function {
               name: encoded_name,
               parameters: main_function_parameter_with_types
                 .into_iter()
@@ -1172,7 +1172,7 @@ fn compile_sources_with_generics_preserved(
             let tparams = tparams_set.iter().sorted().cloned().collect_vec();
             type_lowering_manager.generic_types = tparams_set;
             let main_function_parameter_with_types =
-              vec![(well_known_pstrs::UNDERSCORE_THIS, hir::INT_TYPE)]
+              vec![(well_known_pstrs::UNDERSCORE_THIS, mir::INT_TYPE)]
                 .into_iter()
                 .chain(member.decl.parameters.iter().map(|id| {
                   (
@@ -1196,14 +1196,14 @@ fn compile_sources_with_generics_preserved(
               expression,
               synthetic_functions: mut compiled_functions_to_add,
             } = lower_source_expression(manager, &member.body);
-            let main_fn_type = hir::Type::new_fn_unwrapped(
+            let main_fn_type = mir::Type::new_fn_unwrapped(
               main_function_parameter_with_types.iter().map(|(_, t)| t.clone()).collect_vec(),
               type_lowering_manager.lower_source_type(
                 heap,
                 &type_::Type::from_annotation(&member.decl.type_.return_type),
               ),
             );
-            let original_f = hir::Function {
+            let original_f = mir::Function {
               name: encoded_name,
               parameters: main_function_parameter_with_types
                 .into_iter()
@@ -1225,14 +1225,14 @@ fn compile_sources_with_generics_preserved(
   let SynthesizedTypes { closure_types, mut tuple_types } =
     type_lowering_manager.type_synthesizer.synthesized_types();
   compiled_type_defs.append(&mut tuple_types);
-  compiled_type_defs.push(hir::TypeDefinition {
+  compiled_type_defs.push(mir::TypeDefinition {
     identifier: well_known_pstrs::UNDERSCORE_STR,
     type_parameters: vec![],
     names: vec![],
-    mappings: hir::TypeDefinitionMappings::Enum,
+    mappings: mir::TypeDefinitionMappings::Enum,
   });
 
-  hir::Sources {
+  mir::Sources {
     global_variables: string_manager.all_global_variables(),
     closure_types,
     type_definitions: compiled_type_defs,
@@ -1241,22 +1241,22 @@ fn compile_sources_with_generics_preserved(
   }
 }
 
-fn optimize_by_tail_rec_rewrite(heap: &mut Heap, sources: hir::Sources) -> hir::Sources {
-  let hir::Sources {
+fn optimize_by_tail_rec_rewrite(heap: &mut Heap, sources: mir::Sources) -> mir::Sources {
+  let mir::Sources {
     global_variables,
     closure_types,
     type_definitions,
     main_function_names,
     functions,
   } = sources;
-  hir::Sources {
+  mir::Sources {
     global_variables,
     closure_types,
     type_definitions,
     main_function_names,
     functions: functions
       .into_iter()
-      .map(|f| hir_tail_recursion_rewrite::optimize_function_by_tailrec_rewrite(heap, f))
+      .map(|f| mir_tail_recursion_rewrite::optimize_function_by_tailrec_rewrite(heap, f))
       .collect(),
   }
 }
@@ -1264,11 +1264,11 @@ fn optimize_by_tail_rec_rewrite(heap: &mut Heap, sources: hir::Sources) -> hir::
 pub(crate) fn compile_sources_to_hir(
   heap: &mut Heap,
   sources: &HashMap<ModuleReference, source::Module<Rc<type_::Type>>>,
-) -> hir::Sources {
+) -> mir::Sources {
   let mut sources = compile_sources_with_generics_preserved(heap, sources);
   sources = hir_generics_specialization::perform_generics_specialization(heap, sources);
-  sources = hir_type_deduplication::deduplicate(heap, sources);
-  sources = hir_constant_param_elimination::rewrite_sources(sources);
+  sources = mir_type_deduplication::deduplicate(heap, sources);
+  sources = mir_constant_param_elimination::rewrite_sources(sources);
   sources = optimize_by_tail_rec_rewrite(heap, sources);
   sources
 }
@@ -1277,7 +1277,7 @@ pub(crate) fn compile_sources_to_hir(
 mod tests {
   use crate::{
     ast::{
-      hir,
+      mir,
       source::{self, test_builder, CommentStore, NO_COMMENT_REFERENCE},
       Location, Reason,
     },
@@ -1286,7 +1286,7 @@ mod tests {
     compiler::{
       hir_lowering::ExpressionLoweringManager,
       hir_string_manager::StringManager,
-      hir_type_conversion::{SynthesizedTypes, TypeLoweringManager, TypeSynthesizer},
+      mir_type_conversion::{SynthesizedTypes, TypeLoweringManager, TypeSynthesizer},
     },
   };
   use itertools::Itertools;
@@ -1310,20 +1310,20 @@ mod tests {
     let type_def_mapping = HashMap::from([
       (
         heap.alloc_str_for_test("DUMMY_Foo"),
-        hir::TypeDefinition {
+        mir::TypeDefinition {
           identifier: heap.alloc_str_for_test("DUMMY_Foo"),
           type_parameters: vec![],
           names: vec![],
-          mappings: hir::TypeDefinitionMappings::Struct(vec![hir::INT_TYPE, hir::INT_TYPE]),
+          mappings: mir::TypeDefinitionMappings::Struct(vec![mir::INT_TYPE, mir::INT_TYPE]),
         },
       ),
       (
         heap.alloc_str_for_test("DUMMY_Dummy"),
-        hir::TypeDefinition {
+        mir::TypeDefinition {
           identifier: heap.alloc_str_for_test("DUMMY_Dummy"),
           type_parameters: vec![],
           names: vec![],
-          mappings: hir::TypeDefinitionMappings::Struct(vec![hir::INT_TYPE, hir::INT_TYPE]),
+          mappings: mir::TypeDefinitionMappings::Struct(vec![mir::INT_TYPE, mir::INT_TYPE]),
         },
       ),
     ]);
@@ -1333,19 +1333,19 @@ mod tests {
       vec![
         (
           heap.alloc_str_for_test("_this"),
-          hir::Type::new_id_no_targs(heap.alloc_str_for_test("DUMMY_Dummy")),
+          mir::Type::new_id_no_targs(heap.alloc_str_for_test("DUMMY_Dummy")),
         ),
-        (heap.alloc_str_for_test("foo"), hir::INT_TYPE),
-        (heap.alloc_str_for_test("bar"), hir::INT_TYPE),
+        (heap.alloc_str_for_test("foo"), mir::INT_TYPE),
+        (heap.alloc_str_for_test("bar"), mir::INT_TYPE),
         (
           heap.alloc_str_for_test("closure"),
-          hir::Type::new_id_no_targs(heap.alloc_str_for_test("Closure")),
+          mir::Type::new_id_no_targs(heap.alloc_str_for_test("Closure")),
         ),
         (
           heap.alloc_str_for_test("closure_unit_return"),
-          hir::Type::new_id_no_targs(heap.alloc_str_for_test("Closure")),
+          mir::Type::new_id_no_targs(heap.alloc_str_for_test("Closure")),
         ),
-        (heap.alloc_str_for_test("captured_a"), hir::INT_TYPE),
+        (heap.alloc_str_for_test("captured_a"), mir::INT_TYPE),
       ],
       &type_def_mapping,
       heap,
@@ -1357,7 +1357,7 @@ mod tests {
     let SynthesizedTypes { tuple_types, closure_types } =
       type_lowering_manager.type_synthesizer.synthesized_types();
     let global_variables = string_manager.all_global_variables();
-    let synthetic_module = hir::Sources {
+    let synthetic_module = mir::Sources {
       global_variables,
       closure_types,
       type_definitions: tuple_types,
