+++ crates/samlang-core/src/compiler/lir_unused_name_elimination.rs
@@ -1,5 +1,5 @@
 use crate::{
-  ast::mir::{Expression, Function, GenenalLoopVariable, Sources, Statement, Type},
+  ast::lir::{Expression, Function, GenenalLoopVariable, Sources, Statement, Type},
   common::PStr,
 };
 use itertools::Itertools;
@@ -172,11 +172,11 @@ pub(super) fn optimize_mir_sources_by_eliminating_unused_ones(
 mod tests {
   use crate::{
     ast::{
-      hir::GlobalVariable,
-      mir::{
+      lir::{
         Expression, Function, GenenalLoopVariable, Sources, Statement, Type, TypeDefinition,
         INT_TYPE, ZERO,
       },
+      mir::GlobalVariable,
     },
     Heap,
   };
@@ -255,7 +255,7 @@ mod tests {
               condition: ZERO,
               s1: vec![Statement::binary(
                 heap.alloc_str_for_test(""),
-                crate::ast::hir::Operator::GE,
+                crate::ast::mir::Operator::GE,
                 Expression::Name(heap.alloc_str_for_test("foo"), INT_TYPE),
                 Expression::Name(heap.alloc_str_for_test("bar"), INT_TYPE),
               )],
