+++ crates/samlang-core/src/optimization/optimization_common.rs
@@ -1,5 +1,5 @@
 use crate::{
-  ast::hir::*,
+  ast::mir::*,
   common::{LocalStackedContext, PStr},
 };
